# Swagger/documentation.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api animalerie",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "sandbox-server"
        }
    ],
    "paths": {
...
        "/categorie": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "List all categories",
                "parameters": [
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List all categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Categorie[]",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Categorie"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès à la page refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categorie/{id}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Retourne une categorie",
                "description": "Retourne un categorie",
                "operationId": "287834d7700e96fc8d9e955e1d6d9c48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de la categorie à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne une categorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Achat"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès à la page refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/catégorie/create": {
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Create catégorie",
                "description": "Créer une catégorie",
                "operationId": "639118def56d1866478a89419837f4f1",
                "parameters": [
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Catégorie à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Alimentation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/catégorie/update/{id}": {
            "put": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Update catégorie",
                "description": "Update catégorie in Db",
                "operationId": "f423913961e946597019832e01857fbd",
                "parameters": [
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de l'catégorie à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Catégorie à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Alimentation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/catégorie/delete/{id}": {
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Delete catégorie",
                "description": "Delete catégorie in Db",
                "operationId": "59c300c745b6f70e817ad956a7cc53ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de la catégorie à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression validée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Catégorie supprimé"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Achat": {
                "title": "Achat",
                "properties": {
                    "id": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "id_utilisateur": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "id_produit": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "montatnt": {
                        "nullable": false,
                        "type": "float"
                    },
                    "quantite": {
                        "nullable": false,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Categorie": {
                "title": "Catégorie",
                "properties": {
                    "id": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "nom": {
                        "nullable": false,
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}